bugfix: don't free up page tables more than once
Caused by src/space_mem.cpp:

81     for (unsigned long i = 0; i < 1UL << (o - ord); i++)
82         hpt.update (quota, b + i * (1UL << (ord + PAGE_BITS)), ord, p + i * (1UL << (ord + PAGE_BITS)), Hpt::hw_attr (a), r ? Hpt::TYPE_DN : H    pt::TYPE_UP);
83
84     if (r) {
85
86         for (unsigned i = 0; i < sizeof (loc) / sizeof (*loc); i++)
87             if (loc[i].addr())
88                 loc[i].update (quota, b, o, p, Hpt::hw_attr (a), Hpt::TYPE_DF);
89
90         htlb.merge (cpus);
91     }

during delegate (r==0) the page table in pte.cpp line 96 (below) may get freed,
caused by line 82 (above).

pte.cpp:

92:        if (!e[i].val)
93:            continue;
94:
95:        if (l && !e[i].super())
96:            Pte::destroy(static_cast<P *>(Buddy::phys_to_ptr (e[i].addr())), quota);

Later on during revoke (r!=0 in space_mem.cpp) the very same page table may
get freed multiple times by line 96 (pte.cpp) caused by line 88 (space_mem_cpp).

Original Genode issue genodelabs/genode#1672

+++ src/pte.cpp
@@ -92,6 +92,9 @@ void Pte<P,E,L,B,F>::update (Quota &quota, E v, mword o, E p, mword a, Type t)
         if (!e[i].val)
             continue;
 
+        if (t == TYPE_DF)
+            continue;
+
         if (l && !e[i].super())
             Pte::destroy(static_cast<P *>(Buddy::phys_to_ptr (e[i].addr())), quota);
     }
